SortOption: add UIAlertViewController for sorting and remove the PresentationOptionsView
