+++ VLC-iOS-Bridging-Header.h
@@ -18,3 +18,4 @@
 #import "VLCDownloadViewController.h"
 #import "VLCOpenNetworkStreamViewController.h"
 #import "VLCAboutViewController.h"
+#import "UIColor+Presets.h"
